+++ packages/@leight/file/src/$export/$export.ts
@@ -1,3 +1,6 @@
+export * from "./FileCommitSchema";
 export * from "./FileSchema";
 export * from "./IFile";
+export * from "./IFileCommit";
+export * from "./IFileCommitSchema";
 export * from "./IFileWithPath";