+++ packages/@leight/file/src/$export/IFileCommit.ts
@@ -1,0 +1,1 @@
+export {type IFileCommit} from "../schema/FileCommitSchema";