+++ packages/@leight/file/src/$export/IFileCommitSchema.ts
@@ -1,0 +1,1 @@
+export {type IFileCommitSchema} from "../schema/FileCommitSchema";