+++ packages/@leight/file/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/file
 
+## 0.6.4
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/utils@0.6.14
+
 ## 0.6.3
 
 ### Patch Changes