+++ packages/@leight/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/utils
 
+## 0.6.14
+
+### Patch Changes
+
+- Version bump
+
 ## 0.6.13
 
 ### Patch Changes